Portfolio_Agents_Using_CrewAI/software_engineers/requirements.lock.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
backrefs==5.9
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.2.1
cohere==5.16.1
colorama==0.4.6
coloredlogs==15.0.1
crewai==0.141.0
crewai-tools==0.51.1
cryptography==45.0.5
dataclasses-json==0.6.7
decorator==5.2.1
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
docker==7.1.0
docstring-parser==0.16
durationpy==0.10
embedchain==0.1.128
et-xmlfile==2.0.0
executing==2.2.0
fastapi==0.116.1
fastavro==1.11.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
ghp-import==2.1.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
gptcache==0.1.44
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.4
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.12
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
instructor==1.9.2
ipython==9.4.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
json-repair==0.25.2
json5==0.12.0
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
lancedb==0.24.1
langchain==0.3.26
langchain-cohere==0.3.5
langchain-community==0.3.27
langchain-core==0.3.68
langchain-experimental==0.3.4
langchain-ollama==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.8
langsmith==0.3.45
litellm==1.72.6
mako==1.3.10
markdown==3.8.2
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.114
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.15
mkdocs-material-extensions==1.3.1
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.6.3
mypy-extensions==1.1.0
networkx==3.5
nodeenv==1.9.1
numpy==2.3.1
oauthlib==3.3.1
ollama==0.5.1
onnxruntime==1.22.0
openai==1.95.1
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-instrumentation-asgi==0.56b0
opentelemetry-instrumentation-fastapi==0.56b0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opentelemetry-util-http==0.56b0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
paginate==0.5.7
pandas==2.3.1
parso==0.8.4
pathspec==0.12.1
pdfminer-six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.3.8
portalocker==2.10.1
posthog==3.25.0
pre-commit==4.2.0
prompt-toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
pure-eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyjwt==2.10.1
pymdown-extensions==10.16
pypdf==5.7.0
pypdfium2==4.30.1
pypika==0.48.9
pyproject-hooks==1.2.0
pyreadline3==3.5.4
pyright==1.1.403
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytube==15.0.0
pytz==2025.2
pyvis==0.3.2
pywin32==310
pyyaml==6.0.2
pyyaml-env-tag==1.1
qdrant-client==1.14.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
schema==0.7.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
-e file:///C:/Users/ganes_3ck5/DataScience/Gen_AI/Course_GenAI/Gen_AI_In-Depth/Agentic_AI_Agents/Portfolio_Agents_Using_CrewAI/software_engineers
soupsieve==2.7
sqlalchemy==2.0.41
stack-data==0.6.3
starlette==0.47.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.20.3
tomli==2.2.1
tomli-w==1.2.0
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
types-requests==2.32.4.20250611
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uv==0.7.20
uvicorn==0.35.0
virtualenv==20.31.2
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
